Add IIndicator w/ Identity impl
The IIndicator interface acts as a base for all composable indicators.
This indicator is purposefully symetric w.r.t input/output allowing
easier composition/chaining of indicators with one another.

An event is fired when an indicator updates its output value allowing
data to be pushed through the system.

The IdentityIndicator is the simplest implementation of the IIndcator
interface and represents an indicator that simply uses the input as the
output.