Add IndicatorExtensions and CompositeIndicator
The composite indicator provides a means for combining multiple indicator's
outputs into a single value. This makes the implementations of functionality
like multiplication of indicators easy.

The Of extension method is of particular note, as it allows one to build chains
of indicators, for example, the mean of the std. The std indicator would be the
one that computes its output and then forwards the output to the mean for
averaging.